requirements.txt
anyio==3.7.0
awscli==1.27.155
boto3==1.26.155
botocore==1.29.155
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
dnspython==2.3.0
docutils==0.16
ecdsa==0.18.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi==0.97.0
h11==0.14.0
httptools==0.5.0
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
loguru==0.7.0
MarkupSafe==2.1.3
motor==3.1.2
orjson==3.9.1
packaging==23.1
passlib==1.7.4
pluggy==1.0.0
pyasn1==0.5.0
pydantic==1.10.9
pymongo==4.3.3
pytest==7.3.2
pytest-asyncio==0.21.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==5.4.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.6.3
urllib3==1.26.16
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3